rag-service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic

openai
qdrant-client
